% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectizeInputWithButtons.R
\name{selectizeInputWithValidation}
\alias{selectizeInputWithValidation}
\title{selectizeInput with validation utilities}
\usage{
selectizeInputWithValidation(
  inputId,
  label,
  options = NULL,
  container_id = NULL,
  help_id = NULL,
  label_title = NULL
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{options}{A list of options. See the documentation of \pkg{selectize.js}
for possible options (character option values inside \code{\link[base:AsIs]{base::I()}} will
be treated as literal JavaScript code; see \code{\link[shiny:renderDataTable]{renderDataTable()}}
for details).}

\item{container_id}{id for outer \code{div} (optional). Can use e.g. to add \code{'has-error'} class to style inputs
with \code{shinyjs::addClass}.}

\item{help_id}{id for span with class \code{'help-block'} below input (optional). Can be used for add
error message with \code{shinyjs::html}.}

\item{label_title}{Optional text to appear in info icon tooltip. Used to provide information to user about input.}
}
\description{
selectizeInput with validation utilities
}
\seealso{
\code{\link{selectizeInputWithButtons}}, \code{\link[shinyjs]{html}},
 \code{\link[shinyjs]{addClass}}, \code{\link{toggleSelectizeButtons}}.
}
