% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textInputWithButtons.R
\name{textAreaInputWithButtons}
\alias{textAreaInputWithButtons}
\title{textAreaInput with buttons and validation}
\usage{
textAreaInputWithButtons(
  inputId,
  label,
  ...,
  container_id = NULL,
  help_id = NULL,
  label_title = NULL,
  btn_titletips = TRUE,
  placeholder = ""
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{...}{One or more \code{actionButton}'s to appear to right of input element.}

\item{container_id}{id for outer \code{div} (optional). Can use e.g. to add \code{'has-error'} class to style inputs
with \code{shinyjs::addClass}.}

\item{help_id}{id for span with class \code{'help-block'} below input (optional). Can be used for add
error message with \code{shinyjs::html}.}

\item{label_title}{Optional text to appear in info icon tooltip. Used to provide information to user about input.}

\item{btn_titletips}{if \code{TRUE} (default), converts title attributes of \code{actionButtons} into tooltips.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}
}
\description{
textAreaInput with buttons and validation
}
