% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textInputWithButtons.R
\name{textInputWithValidation}
\alias{textInputWithValidation}
\title{textInput with validation}
\usage{
textInputWithValidation(inputId, label, container_id = NULL, help_id = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{container_id}{id for outer \code{div} (optional). Can use e.g. to add \code{'has-error'} class to style inputs
with \code{shinyjs::addClass}.}

\item{help_id}{id for span with class \code{'help-block'} below input (optional). Can be used for add
error message with \code{shinyjs::html}.}
}
\description{
textInput with validation
}
