% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectizeInputWithButtons.R
\name{toggleSelectizeButtons}
\alias{toggleSelectizeButtons}
\title{Show/hide all buttons in selectizeInputWithButtons}
\usage{
toggleSelectizeButtons(selectize_id, button_ids, condition)
}
\arguments{
\item{selectize_id}{id of \code{selectizeInputWithButtons} element.}

\item{button_ids}{character vector of \code{actionButton} ids associated with \code{selectize_id}.}

\item{condition}{An optional argument to \code{toggleClass}, see 'Details' below.}
}
\description{
When hiding all buttons, some CSS changes are required to allow return of full-width
and rounded corners of selectizeInput.
}
\seealso{
\code{\link{selectizeInputWithButtons}}
}
